Update 05_verifying_feature_pacts.md
Hey, 

After including the `includeWipPactsSince` option in step 1, the local verification task doesn't appear to include the wip pacts. The only way I could pick them up was commenting out the `consumerVersionSelectors`, or explicitly listing the tag. The pending pact was picked up in CI and verified on the broker (use RO token on the provider build locally, and RW in CI)

https://travis-ci.com/github/YOU54F/example-provider/jobs/322285039